Updated privacy policy

+++ terminusgps_tracker/models/customer.py
@@ -1,10 +1,8 @@
 from django.db import models
 from django.contrib.auth import get_user_model
 
-from authorizenet.apicontractsv1 import deleteCustomerProfileRequest
-from authorizenet.apicontrollers import deleteCustomerProfileController
+from django.urls import reverse
 
-from terminusgps_tracker.authorizenetapi.auth import get_merchant_auth
 from terminusgps_tracker.wialonapi.session import WialonSession
 from terminusgps_tracker.wialonapi.items import (
     WialonUser,
@@ -14,12 +12,16 @@ from terminusgps_tracker.wialonapi.items import (
 
 
 class TodoItem(models.Model):
-    label = models.CharField(max_length=128)
-    url = models.URLField()
-    completed = models.BooleanField(default=False)
+    label = models.CharField(max_length=64)
+    view = models.CharField(max_length=512, default="tracker profile")
+    is_complete = models.BooleanField(default=False)
 
     def __str__(self) -> str:
         return self.label
+
+    @property
+    def url(self) -> str:
+        return reverse(self.view)
 
 
 class TodoList(models.Model):
@@ -53,35 +55,25 @@ class TrackerProfile(models.Model):
     def __str__(self) -> str:
         return f"{self.user.username}'s Profile"
 
-    def delete(self, *args, **kwargs):
-        if self.authorizenet_profile_id:
-            self._delete_authorizenet_objects()
-        if all(
-            [
-                self.wialon_user_id,
-                self.wialon_super_user_id,
-                self.wialon_group_id,
-                self.wialon_resource_id,
-            ]
-        ):
-            self._delete_wialon_objects()
-        return super().delete(*args, **kwargs)
-
-    def _delete_authorizenet_objects(self) -> None:
-        if self.authorizenet_profile_id is not None:
-            request = deleteCustomerProfileRequest(
-                merchantAuthentication=get_merchant_auth(),
-                customerProfileId=str(self.authorizenet_profile_id),
-            )
-            controller = deleteCustomerProfileController(request)
-            controller.execute()
-
-    def _delete_wialon_objects(self) -> None:
-        with WialonSession() as session:
-            WialonUser(id=str(self.wialon_user_id), session=session).delete()
-            WialonUnitGroup(id=str(self.wialon_group_id), session=session).delete()
-            WialonResource(id=str(self.wialon_resource_id), session=session).delete()
-            WialonUser(id=str(self.wialon_super_user_id), session=session).delete()
+    def delete(
+        self,
+        delete_wialon_objs: bool = False,
+        delete_authorizenet_objs: bool = False,
+        using=None,
+        keep_parents: bool = False,
+        **kwargs,
+    ) -> tuple[int, dict[str, int]]:
+        if delete_wialon_objs:
+            with WialonSession() as session:
+                WialonUser(id=str(self.wialon_user_id), session=session).delete()
+                WialonUnitGroup(id=str(self.wialon_group_id), session=session).delete()
+                WialonResource(
+                    id=str(self.wialon_resource_id), session=session
+                ).delete()
+                WialonUser(id=str(self.wialon_super_user_id), session=session).delete()
+        if delete_authorizenet_objs:
+            ...
+        return super().delete(using=using, keep_parents=keep_parents, **kwargs)
 
     @property
     def merchantCustomerId(self) -> str:
